refactor: rm Navigator directory
