+++ packages/bezier-react/src/components/NavItem/NavItem.tsx
@@ -8,10 +8,11 @@ import {
   Icon,
   IconSize,
 } from '~/src/components/Icon'
-import styles from '~/src/components/Navigator/Navigator.module.scss'
 import { Text } from '~/src/components/Text'
 
 import type { NavItemProps } from './NavItem.types'
+
+import styles from './NavItem.module.scss'
 
 export const NAV_ITEM_TEST_ID = 'bezier-react-nav-item'
 export const NAV_ITEM_LEFT_ICON_TEST_ID = 'bezier-react-nav-item-left-icon'
