+++ packages/bezier-react/src/index.ts
@@ -63,8 +63,8 @@ export * from '~/src/components/Icon'
 export * from '~/src/components/KeyValueItem'
 export * from '~/src/components/ListItem'
 export * from '~/src/components/Modal'
-export * from '~/src/components/Navigator/NavGroup'
-export * from '~/src/components/Navigator/NavItem'
+export * from '~/src/components/NavGroup'
+export * from '~/src/components/NavItem'
 export * from '~/src/components/OutlineItem'
 export * from '~/src/components/Overlay'
 export * from '~/src/components/ProgressBar'
